requirements.txt
Django==4.2
psycopg2-binary==2.9.7
tenacity==8.2.2
gunicorn==20.1.0
celery[redis]==5.3.4
redis>=4.5.2,<5.0.0 
djangorestframework==3.14.0
djangorestframework-simplejwt==5.4.0
django-cors-headers==3.14.0
whitenoise==6.2.0
requests==2.31.0
Pillow==9.5.0
scikit-learn==1.2.2
drf-yasg==1.21.5
